requirements.txt
flask==3.0.0
tensorflow==2.15.0
numpy
pillow
pandas
gunicorn
